% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmx}
\alias{bmx}
\title{NHANES body measurement data}
\format{
A data frame with 8860 rows and 22 columns.
}
\source{
NHANES
}
\usage{
bmx
}
\description{
A dataset containing NHANES body measurement variables.
}
\keyword{datasets}
