% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{check_bmx_format}
\alias{check_bmx_format}
\title{Validate NHANES BMX dataset format}
\usage{
check_bmx_format(df)
}
\arguments{
\item{df}{data frame to validate}
}
\value{
TRUE if valid, otherwise stops with error
}
\description{
Validate NHANES BMX dataset format
}
