% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{compare}
\alias{compare}
\title{Compare a single ID's value for one variable to the national average (male and female combined)}
\usage{
compare(id, df, name)
}
\arguments{
\item{id}{the ID that we want to search information for (integer)}

\item{df}{the data frame}

\item{name}{name of the variable to be compared with its average (string)}
}
\value{
a string showcasing the comparison. Warning if ID or variable not in the data set
}
\description{
Compare a single ID's value for one variable to the national average (male and female combined)
}
\examples{
df <- get_data()
compare(130378, df, "BMI")
}
