% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_data}
\alias{get_data}
\title{Create a data frame object from the data set}
\usage{
get_data()
}
\value{
data frame with the data set data
}
\description{
Create a data frame object from the data set
}
\examples{
get_data()
}
