% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_info}
\alias{get_info}
\title{Helper function to gather info about a single variable of a specified ID}
\usage{
get_info(id, name, df)
}
\arguments{
\item{id}{the ID that we want to search information for (integer)}

\item{name}{the name of the variable we are gathering the data of (string)}

\item{df}{the data frame}
}
\value{
a string with the value of the chosen variable for the chosen ID and a comment if there is. Warning if no data is found or variable does not exist
}
\description{
Helper function to gather info about a single variable of a specified ID
}
\examples{
df <- get_data()
get_info(130378, "recumbent length", df)
}
