% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_ratio}
\alias{get_ratio}
\title{Get get ratio between two chosen variables for a single ID}
\usage{
get_ratio(id, df, v1, v2)
}
\arguments{
\item{id}{the ID that we want to search information for (integer)}

\item{df}{the data frame}

\item{v1}{name of one value for the ratio (string)}

\item{v2}{name of another value for the ratio (string)}
}
\value{
a string showcasing the information about the desired ratio. Warning if ID or variables not in the data set
}
\description{
Get get ratio between two chosen variables for a single ID
}
\examples{
df <- get_data()
get_ratio(130378, df, "waist circumference", "hip circumference")
}
