% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{health_summary}
\alias{health_summary}
\title{Generate full NHANES summary report}
\usage{
health_summary(df)
}
\arguments{
\item{df}{data frame}
}
\value{
list with metadata, numeric summary, and missingness
}
\description{
Generate full NHANES summary report
}
\examples{
df <- get_data()
health_summary(df)
}
