% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{missing_summary}
\alias{missing_summary}
\title{Summarize missing values in NHANES dataset}
\usage{
missing_summary(df)
}
\arguments{
\item{df}{data frame}
}
\value{
tibble with missing counts and percentages
}
\description{
Summarize missing values in NHANES dataset
}
\examples{
df <- get_data()
missing_summary(df)
}
