% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{nhanes_results}
\alias{nhanes_results}
\title{Get info for all variables or chosen variables for a single ID}
\usage{
nhanes_results(id, df, requests = NULL)
}
\arguments{
\item{id}{the ID that we want to search information for (integer)}

\item{df}{the data frame}

\item{requests}{the names of the variables we want to get info for. If null, get info for all variables (string vector)}
}
\value{
a string with the information for the chosen variables. Warning if ID not found
}
\description{
Get info for all variables or chosen variables for a single ID
}
\examples{
df <- get_data()
nhanes_results(130378, df)
nhanes_results(130378, df, c("recumbent length", "BMI"))
}
