% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot the distribution of a single body measurement}
\usage{
plot_distribution(df, var, bins = 30)
}
\arguments{
\item{df}{from data frame}

\item{var}{name of a column in the df (string)}

\item{bins}{the number of bins for the histogram (default = 30)}
}
\value{
a histogram showing the distribution of a varaible across all rows
}
\description{
Plot the distribution of a single body measurement
}
\examples{
df <- get_data()
plot_distribution(df, "BMI")
}
