% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Density.R
\name{plot_kde}
\alias{plot_kde}
\title{Get KDE Curve with shaded region}
\usage{
plot_kde(df, var, h = NULL)
}
\arguments{
\item{df}{the data frame}

\item{var}{name for the vraible examined (string)}

\item{h}{is bandwidth to control how smooth curve is. Default is set but can be changed by choosing new h.}
}
\value{
a kde curve
}
\description{
Get KDE Curve with shaded region
}
\examples{
df <- get_data()
plot_kde(df, "BMI")
}
