% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Get scatterplot of two variables}
\usage{
plot_scatter(df, x_var, y_var, smooth = TRUE)
}
\arguments{
\item{df}{for the data frame}

\item{x_var}{for one variable to use for x axis (string)}

\item{y_var}{for another variable to use for y axis (string)}

\item{smooth}{curve through the scatterplot showubg trend between data. Can be toggled on and off with TRUE/FALSE (String vector)}
}
\value{
a scatterplot showing trend between two variables
}
\description{
Get scatterplot of two variables
}
\examples{
df <- get_data()
plot_scatter(df, "BMI", "arm circumference")
}
