% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{safe_numeric_summary}
\alias{safe_numeric_summary}
\title{Compute numeric summary statistics}
\usage{
safe_numeric_summary(x, na_rm = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{na_rm}{logical, remove NA values}
}
\value{
tibble with summary statistics
}
\description{
Compute numeric summary statistics
}
