% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{summarize_all}
\alias{summarize_all}
\title{Run full dataset summary}
\usage{
summarize_all(df)
}
\arguments{
\item{df}{data frame}
}
\value{
full summary list
}
\description{
Run full dataset summary
}
\examples{
df <- get_data()
summarize_all(df)
}
