% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{summarize_categorical}
\alias{summarize_categorical}
\title{Summarize categorical NHANES variables}
\usage{
summarize_categorical(df, vars)
}
\arguments{
\item{df}{data frame}

\item{vars}{character vector of variable names}
}
\value{
tibble with counts and proportions
}
\description{
Summarize categorical NHANES variables
}
\examples{
df <- get_data()
summarize_categorical(df, c("head circumference"))
}
