% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{summarize_numeric}
\alias{summarize_numeric}
\title{Summarize numeric NHANES variables}
\usage{
summarize_numeric(df, vars, na_rm = TRUE)
}
\arguments{
\item{df}{data frame}

\item{vars}{character vector of human-readable variable names}

\item{na_rm}{logical, remove NA values}
}
\value{
tibble of summary statistics
}
\description{
Summarize numeric NHANES variables
}
\examples{
df <- get_data()
summarize_numeric(df, c("height", "BMI"))
}
