% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{var_averages}
\alias{var_averages}
\title{Show national average values for the variables in the data frame}
\usage{
var_averages()
}
\value{
a string showcasing the national averages for every variable in the data frame (male and female combined)
}
\description{
Show national average values for the variables in the data frame
}
\examples{
var_averages()
}
