% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{var_names}
\alias{var_names}
\title{Get names for the variables in the data frame}
\usage{
var_names()
}
\value{
a list of strings that represent the human-readable version of the variables in the data frame
}
\description{
Get names for the variables in the data frame
}
\examples{
var_names()
}
